Backend_alchemy/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
pydantic
python-dotenv
pydantic[email]